test: put live Sepolia suite behind //go:build integration
Addresses #690: unit CI no longer compiles or runs live UserOp/Uniswap/sequential tests. Local make test includes -tags=integration so the full suite still runs; make test/unit matches CI. No scheduled workflow.
